Move transform functions to utils/
