+++ src/app/routes/utils/filterEmptyGroupItems/index.js
