+++ src/app/routes/utils/filterEmptyGroupItems/index.test.js
