+++ src/app/routes/utils/filterGroupsWithoutStraplines/index.js
