+++ src/app/routes/utils/filterGroupsWithoutStraplines/index.test.js
